web_app/requirements.txt
dnspython
pymongo
python-dotenv
websockets